+++ docs_src/static_files/basic.py
@@ -1,0 +1,8 @@
+from lilya.apps import Lilya
+from lilya.routing import Include
+from lilya.staticfiles import StaticFiles
+
+
+app = Lilya(routes=[
+    Include('/static', app=StaticFiles(directory='static'), name="static"),
+])